requirements-non-torch.txt
absl-py==2.0.0
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
bottleneck==1.3.5
cachetools==5.3.2
contourpy==1.2.0
cycler==0.12.1
easy-torch==1.3.2
easydict==1.10
einops==0.7.0
filelock==3.9.0
fonttools==4.47.0
fsspec==2023.12.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
grpcio==1.60.0
huggingface-hub==0.19.4
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lightning-utilities==0.10.1
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.0
numpy==1.22.4
oauthlib==3.2.2
packaging==23.1
pandas==1.3.5
Pillow==9.3.0
protobuf==4.23.4
psutil==5.9.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
pytorch-lightning==1.9.4
PyYAML==6.0.1
regex==2023.10.3
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.13.2
sentencepiece==0.2.0
setproctitle==1.3.2
setuptools==59.5.0
sympy==1.10.1
tables==3.7.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
tokenizers==0.15.0
torchmetrics==0.9.3
tqdm==4.66.1
transformers==4.36.2
typing-extensions==4.4.0
urllib3==1.26.13
Werkzeug==3.0.1
yarl==1.9.4
